requirements.txt
transformers
torch
pillow
opencv-python-headless
fastapi
python-multipart
uvicorn